fix: do not sign the request for POST oauth/access_token
Fix Issue #310.

The `POST /oauth/access_token` endpoint does not require the
`Authorization` header to be set [1]. If `signRequired` is `true` in
`makeRequest()`, then OAuth 1.0 parameters will be added to the
`Authorization` header [2,3].  Twitter will then return `401 -
Request Token missing` as a response for the `POST` request.

This might be a bug of Twitter API as Twitter should ignore the
`Authorization` header in this endpoint when it is not required.

[1]: https://developer.twitter.com/en/docs/authentication/api-reference/access_token
[2]: https://github.com/scribejava/scribejava/blob/124745961e999ccede90e2348c6012f8d335eb8a/scribejava-core/src/main/java/com/github/scribejava/core/oauth/OAuth10aService.java#L165
[3]: https://github.com/scribejava/scribejava/blob/124745961e999ccede90e2348c6012f8d335eb8a/scribejava-core/src/main/java/com/github/scribejava/core/oauth/OAuth10aService.java#L86
Signed-off-by: Eng Zer Jun <engzerjun@gmail.com>